requirements.txt
django==4.2.1
django-debug-toolbar==4.1.0
dj-database-url==2.0.0
Pillow==9.5.0
environs[django]==9.5.0
phonenumbers==8.13.13
djangorestframework~=3.14.0
requests~=2.31.0
geopy~=2.3.0
